LPS-124206 Extract methods from DataLayoutBuilder to utils
