+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/actions.es.js
@@ -127,8 +127,9 @@ export const dropFieldSet = ({
 	return {
 		...otherProps,
 		defaultLanguageId: fieldSet.defaultLanguageId,
-		fieldSet: dataLayoutBuilder.getFieldSetDDMForm(fieldSet, {
+		fieldSet: dataLayoutBuilder.getFieldSetDDMForm({
 			availableLanguageIds,
+			fieldSet,
 		}),
 		indexes,
 		parentFieldName,
