+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/components/field-sets/FieldSets.es.js
@@ -21,7 +21,6 @@ import DataLayoutBuilderContext from '../../data-layout-builder/DataLayoutBuilde
 import {DRAG_FIELDSET} from '../../drag-and-drop/dragTypes.es';
 import {containsFieldSet} from '../../utils/dataDefinition.es';
 import {getLocalizedValue} from '../../utils/lang.es';
-import {normalizeDataLayoutRows} from '../../utils/normalizers.es';
 import EmptyState from '../empty-state/EmptyState.es';
 import FieldType from '../field-types/FieldType.es';
 import {getPluralMessage} from './../../utils/lang.es';
@@ -89,10 +88,10 @@ export default function FieldSets({keywords}) {
 		if (fieldSet) {
 			const {context} = appProps;
 			const {defaultDataLayout, id: dataDefinitionId} = fieldSet;
-			const ddmForm = dataLayoutBuilder.getFieldSetDDMForm(
+			const ddmForm = dataLayoutBuilder.getFieldSetDDMForm({
+				availableLanguageIds: dataDefinition.availableLanguageIds,
 				fieldSet,
-				dataDefinition
-			);
+			});
 
 			const [{rows}] = ddmForm.pages;
 
@@ -134,12 +133,6 @@ export default function FieldSets({keywords}) {
 	const deleteFieldSet = useDeleteFieldSet({dataLayoutBuilder});
 	const propagateFieldSet = usePropagateFieldSet();
 
-	const getFieldSet = (fieldSet) => {
-		return dataLayoutBuilder.getFieldSetDDMForm(fieldSet, {
-			availableLanguageIds: dataDefinition.availableLanguageIds,
-		});
-	};
-
 	const onDoubleClick = ({fieldSet}) => {
 		const {activePage, pages} = dataLayoutBuilder.getStore();
 
@@ -175,11 +168,6 @@ export default function FieldSets({keywords}) {
 								fieldSet.defaultLanguageId,
 								fieldSet.name
 							);
-
-							const dataLayoutPages = (
-								fieldSet.defaultDataLayout ||
-								dataLayoutBuilder.getDefaultDataLayout(fieldSet)
-							).dataLayoutPages;
 
 							return (
 								<FieldType
@@ -231,15 +219,11 @@ export default function FieldSets({keywords}) {
 										)
 									}
 									dragType={DRAG_FIELDSET}
-									fieldSet={getFieldSet(fieldSet)}
+									fieldSet={fieldSet}
 									icon="forms"
 									key={fieldSet.dataDefinitionKey}
 									label={fieldSetName}
 									onDoubleClick={onDoubleClick}
-									rows={
-										fieldSet.id &&
-										normalizeDataLayoutRows(dataLayoutPages)
-									}
 								/>
 							);
 						})}
